+++ NEWS.md
@@ -1,5 +1,7 @@
 # modelr 0.1.1.9000
 
+* Added `...` argument to the `typical()` generic function (@jrnold, #42)
+
 # modelr 0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.